pyproject.toml
[project]
name = "test-ss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pillow>=12.1.0",
    "playwright>=1.57.0",
    "uvicorn>=0.40.0",
]
